Move HttpReads into separate files & packages
