+++ src/test/scala/uk/gov/hmrc/play/http/HttpGetSpec.scala
@@ -38,6 +38,7 @@ import play.api.http.HttpVerbs._
 import play.api.libs.json._
 import play.twirl.api.Html
 import uk.gov.hmrc.play.audit.http.HeaderCarrier
+import uk.gov.hmrc.play.http.reads.HtmlHttpReads
 
 import scala.concurrent.Future
