+++ src/test/scala/uk/gov/hmrc/play/http/reads/HtmlReadsSpec.scala
@@ -1,0 +1,36 @@
+/*
+ * Copyright 2015 HM Revenue & Customs
+ *
+ * Licensed under the Apache License, Version 2.0 (the "License");
+ * you may not use this file except in compliance with the License.
+ * You may obtain a copy of the License at
+ *
+ *   http://www.apache.org/licenses/LICENSE-2.0
+ *
+ * Unless required by applicable law or agreed to in writing, software
+ * distributed under the License is distributed on an "AS IS" BASIS,
+ * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+ * See the License for the specific language governing permissions and
+ * limitations under the License.
+ */
+
+package uk.gov.hmrc.play.http.reads
+
+import play.twirl.api.Html
+import uk.gov.hmrc.play.http.HttpResponse
+
+class HtmlReadsSpec extends HttpReadsSpec {
+  "HtmlHttpReads" should {
+    "convert a successful response body to HTML" in {
+      val reads = new HtmlHttpReads with StubThatReturnsTheResponse
+
+      reads.readToHtml.read(exampleVerb, exampleUrl, HttpResponse(0, responseString = Some("<p>hello</p>"))) should (
+        be (an[Html]) and have('text("<p>hello</p>"))
+      )
+    }
+    "pass through any failure" in {
+      val reads = new HtmlHttpReads with StubThatThrowsAnException
+      an[Exception] should be thrownBy reads.readToHtml.read(exampleVerb, exampleUrl, exampleResponse)
+    }
+  }
+}
